# docs/openapi.yaml
openapi: 3.1.0
info:
  title: TechNest Shop API
  version: 1.0.0
  description: >
    Versioned OpenAPI contract for the current TechNest Shop backend. This file is
    the authoritative machine-readable API contract; README keeps the operational
    snapshot and docs/API_CONTRACT.md keeps a concise human summary.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: /
...
paths:
...
  /api/orders:
    post:
      tags: [Orders]
      operationId: createOrder
      summary: Create order
      description: Authenticated customer flow. Business rules enforce stock locking and ownership.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderRequest'
      responses:
        '201':
          description: Order created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDetailResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
    get:
      tags: [Orders]
      operationId: listAdminOrders
      summary: Search all orders
      description: Staff or admin only.
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            $ref: '#/components/schemas/OrderStatus'
        - in: query
          name: from
          schema:
            type: string
            format: date
        - in: query
          name: to
          schema:
            type: string
            format: date
        - in: query
          name: q
          schema:
            type: string
      responses:
        '200':
          description: Order list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderSummaryResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/orders/me:
    get:
      tags: [Orders]
      operationId: listMyOrders
      summary: List current user's orders
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user's order summaries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderSummaryResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/orders/{id}:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    get:
      tags: [Orders]
      operationId: getOrder
      summary: Get order detail
      description: Owner or staff/admin.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Order detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDetailResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/orders/{id}/status:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    put:
      tags: [Orders]
      operationId: updateOrderStatus
      summary: Update order status or payment status
      description: >
        Staff/admin can manage shipping and payment transitions. Customers can only
        mark their own shipping orders as delivered.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderStatusUpdateRequest'
      responses:
        '200':
          description: Updated order detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDetailResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/admin/users/{id}:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    get:
      tags: [Admin Users]
      operationId: getUser
      summary: Get user by id
      description: Admin only.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Admin Users]
      operationId: updateUser
      summary: Update user
      description: Admin only.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
    delete:
      tags: [Admin Users]
      operationId: deleteUser
      summary: Delete user
      description: Admin only.
      security:
        - bearerAuth: []
      responses:
        '204':
          description: User deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdPath:
      in: path
      name: id
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
...
  responses:
    BadRequest:
      description: Request validation failed or payload is invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    Unauthorized:
      description: Authentication is required or credentials are invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    Forbidden:
      description: Authenticated but not allowed to perform this action
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    Conflict:
      description: Request conflicts with existing data or business rules
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
  schemas:
    ApiErrorResponse:
      type: object
      properties:
        message:
          type: string
        status:
          type: integer
        error:
          type: string
        validationErrors:
          type: object
          additionalProperties:
            type: string
      required: [message, status, error]
...
    OrderStatus:
      type: string
      enum: [PENDING, SHIPPING, DELIVERED]
    PaymentStatus:
      type: string
      enum: [UNPAID, PAID, FAILED, REFUNDED]
    PaymentMethod:
      type: string
      enum: [cod, bank, card]
    OrderItemRequest:
      type: object
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
        qty:
          type: integer
          minimum: 1
      required: [id, qty]
    OrderRequest:
      type: object
      properties:
        items:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/OrderItemRequest'
        address:
          type: array
          minItems: 5
          maxItems: 5
          items:
            type: string
            maxLength: 120
        payment:
          $ref: '#/components/schemas/PaymentMethod'
      required: [items, address, payment]
    OrderItemResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        qty:
          type: integer
        price:
          type: number
      required: [id, name, qty, price]
    OrderSummaryResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        orderNumber:
          type: string
        customerName:
          type: string
        customerEmail:
          type: string
          format: email
        itemCount:
          type: integer
        subtotal:
          type: number
        shipping:
          type: number
        total:
          type: number
        paymentMethod:
          type: string
        status:
          $ref: '#/components/schemas/OrderStatus'
        paymentStatus:
          $ref: '#/components/schemas/PaymentStatus'
        placedAt:
          type: string
      required:
        - id
        - orderNumber
        - customerName
        - customerEmail
        - itemCount
        - subtotal
        - shipping
        - total
        - paymentMethod
        - status
        - paymentStatus
        - placedAt
    OrderDetailResponse:
      allOf:
        - $ref: '#/components/schemas/OrderSummaryResponse'
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/OrderItemResponse'
          required: [items]
    OrderStatusUpdateRequest:
      type: object
      properties:
        status:
          $ref: '#/components/schemas/OrderStatus'
        paymentStatus:
          $ref: '#/components/schemas/PaymentStatus'
      anyOf:
        - required: [status]
        - required: [paymentStatus]
...
    UserRole:
      type: string
      enum: [ADMIN, STAFF, CUSTOMER, admin, staff, customer]
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
          format: email
        username:
          type: string
        fullName:
          type: string
        phone:
          type: string
        addressText:
          type: string
        avatarUrl:
          type: string
        role:
          type: string
      required: [id, email, username, role]
...
    UpdateUserRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          maxLength: 255
        username:
          type: string
          pattern: '^\\s*$|^[A-Za-z0-9._-]{3,50}$'
          maxLength: 50
        password:
          type: string
          minLength: 6
          maxLength: 100
        fullName:
          type: string
          maxLength: 120
        phone:
          type: string
          pattern: '^\\s*$|^[0-9+()\\- ]{8,20}$'
        addressText:
          type: string
          maxLength: 255
        role:
          $ref: '#/components/schemas/UserRole'
...